discord/blot/requirements.txt
pynacl
flask
webcolors
